+++ config/constants.go
@@ -4,4 +4,4 @@ package config
 const ProgramName = "limo"
 
 // Version is the semver-compliant program version
-const Version = "0.4.0"
+const Version = "0.5.0"