+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
+### Added
+- Show URL when listing stars
 
 ## [0.4.0]
 ### Added