+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
+
+## [0.5.0]
 ### Added
 - Add `prune` command to prune unstarred repos
 - Add `-b/--browse` flag to `list` command to open items in browser
@@ -47,7 +49,8 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 ## 0.1.0 - 2016-06-27
 ### Initial Release
 
-[Unreleased]: https://github.com/hoop33/limo/compare/v0.4.0...HEAD
+[Unreleased]: https://github.com/hoop33/limo/compare/v0.5.0...HEAD
+[0.5.0]: https://github.com/hoop33/limo/compare/v0.4.0...v0.5.0
 [0.4.0]: https://github.com/hoop33/limo/compare/v0.3.1...v0.4.0
 [0.3.1]: https://github.com/hoop33/limo/compare/v0.3.0...v0.3.1
 [0.3.0]: https://github.com/hoop33/limo/compare/v0.2.0...v0.3.0